Major refactor in persuit of perfect transforms
So it looks like matrices are out, and in are a completely different way
of handling everyting.

On the flip side, it looks like this is working pretty well.

We should assume if no from is provided - take the given state of the
element when the animation starts.

Also, I think we should wait to calculate the from frame when the tween
is in the queue. This will simplify the tweenElement code a bit.  This
way we can chain various states together very easily.