Fixed a major potential race condition on .toRemove
Because I queued up removel for the next frame, there's the possibility
that a new animation is added between the end of the final frame and and
the start of the next frame when removal would begin. Now removal
happens at the end of the frame, and will ensure due to the single
threaded nature of JS, that there never can be that race condition.