Reworking API a bit
Moving this lib in a new direction.  The whole concept of animating a
single element and making that the queue, sounded attractive at first,
however it breaks down pretty quickly if you want to do anything more
technical.  Instead with the new queue based direction, it made sense to
rename the methods to not reference an element but rather be a bit more
agnostic and enable the ability to queue up animations/tweens/springs
without kicking off the animation sequence